Score tracking bug fixes

+++ client/android/app/src/main/java/com/lucive/managers/LocalStorageManager.java
@@ -20,6 +20,7 @@ import java.util.HashSet;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Set;
+import java.util.concurrent.atomic.AtomicLong;
 
 public class LocalStorageManager {
     private static LocalStorageManager instance;
@@ -29,16 +30,26 @@ public class LocalStorageManager {
     private final Set<LocalStorageObserver> observers = new HashSet<>();
     private final HeartbeatDAO heartbeatDAO;
     private final DeviceStatusDAO deviceStatusDAO;
+    private final AtomicLong lastHeartbeatCleanup = new AtomicLong(0);
+    private final AtomicLong lastDeviceStatusCleanup = new AtomicLong(0);
 
     public static final String WORDS_KEY = "words";
     public static final String RULES_KEY = "rules";
     public static final String USER_KEY = "user";
+    public static final String LAST_HEARTBEAT_CLEANUP_KEY = "last_heartbeat_cleanup";
+    public static final String LAST_DEVICE_STATUS_CLEANUP_KEY = "last_device_status_cleanup";
+    public static final long DAYS_10 = 10 * 24 * 60 * 60;
 
     private LocalStorageManager(Context context) {
         this.sharedPreferences = context.getSharedPreferences("MyPreferences", Context.MODE_PRIVATE);
         this.gson = new Gson();
         this.heartbeatDAO = new HeartbeatDAO(context);
         this.deviceStatusDAO = new DeviceStatusDAO(context);
+
+        long lastHeartbeatCleanupTime = sharedPreferences.getLong(LAST_HEARTBEAT_CLEANUP_KEY, System.currentTimeMillis() / 1000);
+        lastHeartbeatCleanup.set(lastHeartbeatCleanupTime);
+        long lastDeviceStatusCleanupTime = sharedPreferences.getLong(LAST_DEVICE_STATUS_CLEANUP_KEY, System.currentTimeMillis() / 1000);
+        lastDeviceStatusCleanup.set(lastDeviceStatusCleanupTime);
     }
 
     public static synchronized LocalStorageManager getInstance(Context context) {
@@ -107,8 +118,13 @@ public class LocalStorageManager {
     }
 
     private void cleanOldDeviceStatuses() {
+        if (System.currentTimeMillis() / 1000 - lastDeviceStatusCleanup.get() < DAYS_10) {
+            return;
+        }
         long cutoffTimeSeconds = AppUtils.getDayStartNDaysBefore(10) / 1000;
         deviceStatusDAO.deleteOldDeviceStatuses(cutoffTimeSeconds);
+        lastDeviceStatusCleanup.set(System.currentTimeMillis() / 1000);
+        sharedPreferences.edit().putLong(LAST_DEVICE_STATUS_CLEANUP_KEY, lastDeviceStatusCleanup.get()).apply();
     }
 
     public List<DeviceStatus> getDeviceStatuses(final String date) {
@@ -124,7 +140,6 @@ public class LocalStorageManager {
                 filteredDeviceStatuses.add(deviceStatus);
             }
         }
-
         return filteredDeviceStatuses;
     }
 
@@ -155,8 +170,13 @@ public class LocalStorageManager {
     }
 
     private void cleanOldHeartbeats() {
+        if (System.currentTimeMillis() / 1000 - lastHeartbeatCleanup.get() < DAYS_10) {
+            return;
+        }
         long cutoffTimeSeconds = AppUtils.getDayStartNDaysBefore(10) / 1000;
         heartbeatDAO.deleteOldHeartbeats(cutoffTimeSeconds);
+        lastHeartbeatCleanup.set(System.currentTimeMillis() / 1000);
+        sharedPreferences.edit().putLong(LAST_HEARTBEAT_CLEANUP_KEY, lastHeartbeatCleanup.get()).apply();
     }
 
     public void setUser(final User user) {
